# internal/apidoc/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: Aegis
    url: https://github.com/wisonwang/Aegis
  description: Aegis — AI Data Supply Gateway. 受治理的 DataAPI + 面向 AI Agent 的 MCP 服务。默认拒绝：每条查询都经过权限引擎重写与脱敏。
  license:
    name: MIT
  termsOfService: https://github.com/wisonwang/Aegis
  title: Aegis DataAPI
  version: "1.0"
paths:
...
  /admin/api/datasets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Datasets
      tags:
      - datasets
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Create Dataset
      tags:
      - datasets
  /admin/api/datasets/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Dataset
      tags:
      - datasets
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Get Dataset
      tags:
      - datasets
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Update Dataset
      tags:
      - datasets
  /admin/api/datasets/{id}/masks:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Dataset Masks
      tags:
      - datasets
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Upsert Dataset Mask
      tags:
      - datasets
  /admin/api/datasets/{id}/masks/{mask}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: mask
        in: path
        name: mask
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Dataset Mask
      tags:
      - datasets
  /admin/api/datasets/{id}/permissions:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Dataset Permissions
      tags:
      - datasets
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Create Dataset Permission
      tags:
      - datasets
  /admin/api/datasets/{id}/permissions/{perm}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: perm
        in: path
        name: perm
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Dataset Permission
      tags:
      - datasets
  /admin/api/datasets/{id}/policies:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Dataset Policies
      tags:
      - datasets
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Create Dataset Policy
      tags:
      - datasets
  /admin/api/datasets/{id}/policies/{policy}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: policy
        in: path
        name: policy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Dataset Policy
      tags:
      - datasets
  /admin/api/datasets/{id}/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Publish Dataset
      tags:
      - datasets
  /admin/api/datasets/{id}/semantics:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Dataset Semantics
      tags:
      - datasets
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Upsert Dataset Semantic
      tags:
      - datasets
  /admin/api/datasets/{id}/semantics/{sem}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: sem
        in: path
        name: sem
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Dataset Semantic
      tags:
      - datasets
  /admin/api/datasets/{id}/unpublish:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Unpublish Dataset
      tags:
      - datasets
...
  /api/v1/datasets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: list Datasets
      tags:
      - datasets
  /api/v1/datasets/{id}:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: get Dataset
      tags:
      - datasets
  /api/v1/datasets/{id}/query:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: query Dataset
      tags:
      - datasets
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT> — 通过 POST /api/v1/login 获取"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"